tools/transaction-tester/requirements.txt
streamlit==1.44.1
kafka-python==2.2.15
requests==2.32.3
PyYAML==6.0.2
pytest==8.3.5

